Moves comaparator to own folder
