+++ lib/src/comparator/string_comparator.dart
