+++ lib/src/function/date_time_function.dart
@@ -1,4 +1,4 @@
-import 'package:boolean_rhapsody/src/date_time_comparator.dart';
+import 'package:boolean_rhapsody/src/comparator/date_time_comparator.dart';
 import '../evaluation_context.dart';
 import '../fuzzy_boolean.dart';
 import 'rule_function.dart';
