+++ lib/src/function/set_function.dart
@@ -1,7 +1,7 @@
 import '../evaluation_context.dart';
 import '../fuzzy_boolean.dart';
 import 'rule_function.dart';
-import '../set_comparator.dart';
+import '../comparator/set_comparator.dart';
 
 /// A function that evaluates a boolean condition based on set operations.
 ///
