+++ test/comparator/date_time_comparator_test.dart
@@ -1,4 +1,4 @@
-import 'package:boolean_rhapsody/src/date_time_comparator.dart';
+import 'package:boolean_rhapsody/src/comparator/date_time_comparator.dart';
 import 'package:test/test.dart';
 
 void main() {
