+++ test/function/list_size_function_test.dart
@@ -1,6 +1,6 @@
 import 'package:boolean_rhapsody/boolean_rhapsody.dart';
 import 'package:boolean_rhapsody/src/function/list_size_function.dart';
-import 'package:boolean_rhapsody/src/number_comparator.dart';
+import 'package:boolean_rhapsody/src/comparator/number_comparator.dart';
 import 'package:test/test.dart';
 
 const defaultPrefixes = ["c", "v", "p", "d"];
